Implement new Account (& Query) init logic 🎉🚀
