+++ packages/stores/src/account/types.ts
@@ -1,10 +1,4 @@
-import { DeepReadonly } from "utility-types";
 import { MsgOpt } from "@keplr-wallet/stores";
-import { OsmosisAccount } from "./osmosis";
-
-export interface HasOsmosisAccount {
-  osmosis: DeepReadonly<OsmosisAccount>;
-}
 
 export interface OsmosisMsgOpts {
   readonly createPool: MsgOpt;
@@ -23,3 +17,47 @@ export interface OsmosisMsgOpts {
   readonly beginUnlocking: MsgOpt;
   readonly unlockPeriodLock: MsgOpt;
 }
+
+export const defaultMsgOpts: OsmosisMsgOpts = {
+  createPool: {
+    type: "osmosis/gamm/create-pool",
+    gas: 250000,
+  },
+  joinPool: {
+    type: "osmosis/gamm/join-pool",
+    gas: 140000,
+    shareCoinDecimals: 18,
+  },
+  joinSwapExternAmountIn: {
+    type: "osmosis/gamm/join-swap-extern-amount-in",
+    gas: 140000,
+    shareCoinDecimals: 18,
+  },
+  exitPool: {
+    type: "osmosis/gamm/exit-pool",
+    gas: 140000,
+    shareCoinDecimals: 18,
+  },
+  swapExactAmountIn: {
+    type: "osmosis/gamm/swap-exact-amount-in",
+    gas: 250000,
+  },
+  swapExactAmountOut: {
+    type: "osmosis/gamm/swap-exact-amount-out",
+    gas: 250000,
+  },
+  lockTokens: {
+    type: "osmosis/lockup/lock-tokens",
+    gas: 250000,
+  },
+  beginUnlocking: {
+    type: "osmosis/lockup/begin-unlock-period-lock",
+    // Gas per msg
+    gas: 140000,
+  },
+  unlockPeriodLock: {
+    type: "osmosis/lockup/unlock-period-lock",
+    // Gas per msg
+    gas: 140000,
+  },
+};
