+++ packages/stores/src/queries/index.ts
@@ -8,4 +8,3 @@ export * from "./pool-share";
 export * from "./pools";
 
 export * from "./store";
-export * from "./types";
