--- packages/stores/src/queries/types.ts
@@ -1,5 +1,0 @@
-import { QueriesOsmosis } from ".";
-
-export interface HasOsmosisQueries {
-  osmosis: QueriesOsmosis;
-}
