+++ packages/web/hooks/use-amount-config.ts
@@ -1,26 +1,25 @@
 import { useMemo } from "react";
 import { ObservableAmountConfig } from "@osmosis-labs/stores";
-import { ChainGetter, ObservableQueryBalances } from "@keplr-wallet/stores";
+import { ChainGetter, IQueriesStore } from "@keplr-wallet/stores";
 import { AppCurrency } from "@keplr-wallet/types";
 
 export const useAmountConfig = (
   chainGetter: ChainGetter,
+  queriesStore: IQueriesStore,
   chainId: string,
   sender: string,
-  currency: AppCurrency,
-  queryBalances: ObservableQueryBalances
+  currency: AppCurrency
 ) =>
   useMemo(() => {
     const config = new ObservableAmountConfig(
       chainGetter,
+      queriesStore,
       chainId,
       sender,
-      currency,
-      queryBalances
+      currency
     );
     config.setChain(chainId);
-    config.setQueryBalances(queryBalances);
     config.setSender(sender);
     config.setCurrency(currency);
     return config;
-  }, [chainGetter, chainId, sender, currency, queryBalances]);
+  }, [chainGetter, queriesStore, chainId, sender, currency]);
