+++ packages/web/pages/pool/[id].tsx
@@ -20,15 +20,14 @@ import { useStore } from "../../stores";
 
 const Pool: FunctionComponent = observer(() => {
   const router = useRouter();
-  const { chainStore, queriesOsmosisStore, accountStore, priceStore } =
-    useStore();
+  const { chainStore, queriesStore, accountStore, priceStore } = useStore();
 
   const { id: poolId } = router.query;
   const { chainId } = chainStore.osmosis;
 
-  const queries = queriesOsmosisStore.get(chainId);
+  const queryOsmosis = queriesStore.get(chainId).osmosis;
   const account = accountStore.getAccount(chainStore.osmosis.chainId);
-  const pool = queries.queryGammPools.getPool(poolId as string);
+  const pool = queryOsmosis.queryGammPools.getPool(poolId as string);
   const { bech32Address } = accountStore.getAccount(chainId);
   const fiat = priceStore.getFiatCurrency(priceStore.defaultVsCurrency)!;
 
@@ -68,10 +67,13 @@ const Pool: FunctionComponent = observer(() => {
   if (pool) {
     totalValueLocked = pool.computeTotalValueLocked(priceStore);
     userLockedValue = totalValueLocked?.mul(
-      queries.queryGammPoolShare.getAllGammShareRatio(bech32Address, pool.id)
+      queryOsmosis.queryGammPoolShare.getAllGammShareRatio(
+        bech32Address,
+        pool.id
+      )
     );
     userBondedValue = totalValueLocked
-      ? queries.queryGammPoolShare.getLockedGammShareValue(
+      ? queryOsmosis.queryGammPoolShare.getLockedGammShareValue(
           bech32Address,
           pool.id,
           totalValueLocked,
@@ -80,7 +82,7 @@ const Pool: FunctionComponent = observer(() => {
       : undefined;
     userAvailableValue = !pool.totalShare.toDec().equals(new Dec(0))
       ? totalValueLocked.mul(
-          queries.queryGammPoolShare
+          queryOsmosis.queryGammPoolShare
             .getAvailableGammShare(bech32Address, pool.id)
             .quo(pool.totalShare)
         )
@@ -89,7 +91,7 @@ const Pool: FunctionComponent = observer(() => {
       ratio: new RatePretty(asset.weight.quo(pool.totalWeight)),
       asset: asset.amount
         .mul(
-          queries.queryGammPoolShare.getAllGammShareRatio(
+          queryOsmosis.queryGammPoolShare.getAllGammShareRatio(
             bech32Address,
             pool.id
           )
@@ -97,19 +99,19 @@ const Pool: FunctionComponent = observer(() => {
         .trim(true)
         .shrink(true),
     }));
-    userLockedAssets = queries.queryGammPoolShare
+    userLockedAssets = queryOsmosis.queryGammPoolShare
       .getShareLockedAssets(
         bech32Address,
         pool.id,
-        queries.queryLockableDurations.lockableDurations
+        queryOsmosis.queryLockableDurations.lockableDurations
       )
       .map((lockedAsset) =>
         // calculate APR% for this pool asset
         ({
           ...lockedAsset,
-          apr: queries.queryIncentivizedPools.isIncentivized(pool.id)
+          apr: queryOsmosis.queryIncentivizedPools.isIncentivized(pool.id)
             ? new RatePretty(
-                queries.queryIncentivizedPools.computeAPY(
+                queryOsmosis.queryIncentivizedPools.computeAPY(
                   pool.id,
                   lockedAsset.duration,
                   priceStore,
@@ -121,7 +123,7 @@ const Pool: FunctionComponent = observer(() => {
       );
     externalGuages = (ExternalIncentiveGaugeAllowList[pool.id] ?? []).map(
       ({ gaugeId, denom }) => {
-        const observableGauge = queries.queryGauge.get(gaugeId);
+        const observableGauge = queryOsmosis.queryGauge.get(gaugeId);
         const currency = chainStore
           .getChain(chainStore.osmosis.chainId)
           .findCurrency(denom);
@@ -138,13 +140,14 @@ const Pool: FunctionComponent = observer(() => {
         };
       }
     );
-    guages = queries.queryLockableDurations.lockableDurations.map(
+    guages = queryOsmosis.queryLockableDurations.lockableDurations.map(
       (duration) => {
-        const guageId = queries.queryIncentivizedPools.getIncentivizedGaugeId(
-          pool.id,
-          duration
-        )!;
-        return queries.queryGauge.get(guageId);
+        const guageId =
+          queryOsmosis.queryIncentivizedPools.getIncentivizedGaugeId(
+            pool.id,
+            duration
+          )!;
+        return queryOsmosis.queryGauge.get(guageId);
       }
     );
 
@@ -167,7 +170,7 @@ const Pool: FunctionComponent = observer(() => {
   // eject to pools page if pool does not exist
   useEffect(() => {
     return autorun(() => {
-      if (queries.queryGammPools.poolExists(poolId as string) === false) {
+      if (queryOsmosis.queryGammPools.poolExists(poolId as string) === false) {
         router.push("/pools");
       }
     });
@@ -267,7 +270,7 @@ const Pool: FunctionComponent = observer(() => {
           </div>
           {pool &&
             guages &&
-            queries.queryIncentivizedPools.isIncentivized(pool.id) && (
+            queryOsmosis.queryIncentivizedPools.isIncentivized(pool.id) && (
               <>
                 <div className="flex gap-9 place-content-between pt-10">
                   {externalGuages?.map(
@@ -292,7 +295,7 @@ const Pool: FunctionComponent = observer(() => {
                     <PoolGaugeCard
                       key={i}
                       days={guage.lockupDuration.humanize()}
-                      apr={queries.queryIncentivizedPools
+                      apr={queryOsmosis.queryIncentivizedPools
                         .computeAPY(
                           pool.id,
                           guage.lockupDuration,
@@ -303,7 +306,7 @@ const Pool: FunctionComponent = observer(() => {
                         .toString()}
                       isLoading={
                         guage.isFetching ||
-                        queries.queryIncentivizedPools.isAprFetching
+                        queryOsmosis.queryIncentivizedPools.isAprFetching
                       }
                       isSuperfluid={
                         superfluid &&
